requirements.txt
pywebview==5.4
streamlit==1.44.0
